third_party/WebKit/LayoutTests/bluetooth/server/getPrimaryService/gen-garbage-collection-ran-during-success.html

<!-- Generated by //third_party/WebKit/LayoutTests/bluetooth/generate.py -->
<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script src="../../../resources/bluetooth/web-bluetooth-test.js"></script>
<script src="../../../resources/mojo-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  let promise;
  return getHealthThermometerDevice({
      filters: [{services: ['health_thermometer']}]})
    .then(([device]) => {
      promise = assert_promise_rejects_with_message(
        device.gatt.getPrimaryService('health_thermometer'),
        new DOMException(
          'GATT Server is disconnected. Cannot retrieve services. ' +
          '(Re)connect first with `device.gatt.connect`.',
          'NetworkError'));
      device.gatt.disconnect();
    })
    .then(runGarbageCollection)
    .then(() => promise);
}, 'Garbage Collection ran during a getPrimaryService call that succeeds. ' +
   'Should not crash.');

</script>
